+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `generateMonotonic()` and `generateCoreMonotonic()` to `Scru128Generator`
 
+### Deprecated
+
+- `Scru128Generator#lastStatus` and `Scru128Generator.Status`
+
 ## v2.3.0 - 2023-02-13
 
 ### Fixed